Upgrade kamu cli version (#104)

+++ src/app/api-server/src/config.rs
@@ -204,7 +204,7 @@ pub struct RemoteDatabaseConfig {
     pub credentials_policy: DatabaseCredentialsPolicyConfig,
     pub database_name: String,
     pub host: String,
-    pub port: Option<u32>,
+    pub port: Option<u16>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
